Refactored the notebooks
Got rid of most of the cell outputs, as well as a lot of comments I made
along the way that helped me follow my thought process from one time to
the other. These comments are either useless (because the problem has
been solved) or have been reported in the paper.